requirements.txt
asgiref==3.3.4
attrs==21.2.0
base58==1.0.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
conseil==0.1.4
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-appconf==1.0.4
django-compressor==2.4.1
django-compressor-toolkit==0.6.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-decouple==2.1
django-extensions==3.1.3
django-heroku==0.3.1
django-model-utils==4.1.1
django-picklefield==3.0.1
djangorestframework==3.12.4
djcall==0.4.6
drf-jwt==1.19.1
ecdsa==0.17.0
fastecdsa==2.1.5
filelock==3.0.12
fire==0.4.0
gunicorn==20.1.0
huggingface-hub==0.0.8
humanize==3.9.0
idna==2.10
iniconfig==1.1.1
joblib==1.0.1
jsonlight==0.0.5
loguru==0.5.3
mnemonic==0.19
mysqlclient==2.0.3
netstruct==1.1.2
nltk==3.6.2
nodejs==0.1.1
npm==0.1.1
numpy==1.20.3
oauthlib==3.1.1
optional-django==0.1.0
packaging==20.9
pendulum==2.1.2
Pillow==8.2.0
pluggy==0.13.1
ply==3.11
psycopg2==2.9.1
py==1.10.0
pyasn1==0.4.8
pyblake2==1.1.2
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pysodium==0.7.9
pytest==6.2.4
pytest-cov==2.12.1
pytezos==2.2.8
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
qrcode==6.1
rcssmin==1.0.6
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.7.2
sacremoses==0.0.45
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tenacity==7.0.0
termcolor==1.1.0
tokenizers==0.10.3
toml==0.10.2
tqdm==4.61.0
transformers==4.6.1
urllib3==1.26.5
whitenoise==5.2.0
